Convert RazorProjectService to an interface
Just converting another pure abstract class to an interface.
