+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/DocumentSynchronization/DocumentDidCloseEndpoint.cs
@@ -16,13 +16,13 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.DocumentSynchronization;
 internal class DocumentDidCloseEndpoint : IRazorNotificationHandler<DidCloseTextDocumentParams>, ITextDocumentIdentifierHandler<DidCloseTextDocumentParams, TextDocumentIdentifier>
 {
     private readonly ProjectSnapshotManagerDispatcher _projectSnapshotManagerDispatcher;
-    private readonly RazorProjectService _projectService;
+    private readonly IRazorProjectService _projectService;
 
     public bool MutatesSolutionState => true;
 
     public DocumentDidCloseEndpoint(
         ProjectSnapshotManagerDispatcher projectSnapshotManagerDispatcher,
-        RazorProjectService projectService)
+        IRazorProjectService projectService)
     {
         _projectSnapshotManagerDispatcher = projectSnapshotManagerDispatcher ?? throw new ArgumentNullException(nameof(projectSnapshotManagerDispatcher));
         _projectService = projectService ?? throw new ArgumentNullException(nameof(projectService));
